[v6] ramips: add additional support for MQmaker WiTi board
The board is based on mt7621AT cpu, and has 16mb nor flash, 256mb of ram, 2 sata ports, microsd card slot, 1 USB 3.0 port and at least one 2.4 and one 5 ghz antenna.
This is the 6th submission that adds support for XHCI in the device tree file, along with switching the location of the 2 radio's and addition of the kmod-i2c-mt7621 in the default packages of the profile.
Patch v1 added initial support for MQmaker WiTi board. The device tree is based on PBR-M1.
Patch v2 changed the flash chip ID (w25q256 to gd25q128).
Patch v3 fixed the left-out entry for WiTi in 02_network.
Patch v4 fixed the naming scheme for MQmaker WiTi board.
Patch v5 reverts Patch v2 and introduces the use of 'jedec,spi-nor' to detect the flash chip ID.

Signed-off-by: Sebastian Careba <nitroshift@yahoo.com>